src/requirements.txt
django~=4.2.3
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
pymysql~=1.1.0